+++ src/components/BurgerIngredients/BurgerIngredients.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef, useState } from 'react';
-import PropTypes, { string } from 'prop-types';
 import { Tab } from '@ya.praktikum/react-developer-burger-ui-components';
 import { useSelector } from 'react-redux';
 import { useInView } from 'react-intersection-observer';
