+++ src/utils/cookieUtils.ts
@@ -1,4 +1,4 @@
-export function getCookie(name) {
+export function getCookie(name: string): string | undefined {
 	const matches = document.cookie.match(
 		new RegExp(
 			'(?:^|; )' +
@@ -10,7 +10,7 @@ export function getCookie(name) {
 	return matches ? decodeURIComponent(matches[1]) : undefined;
 }
 
-export function setCookie(name, value, props) {
+export function setCookie(name: string, value: string, props: { [key: string]: string | number | Date | boolean } = {}) {
 	props = {
 		path: '/',
 		...props,
@@ -38,6 +38,6 @@ export function setCookie(name, value, props) {
 	document.cookie = updatedCookie;
 }
 
-export function deleteCookie(name) {
+export function deleteCookie(name: string) {
 	setCookie(name, '', { expires: -1 });
 }
